Feat: 최근 검색 기록 컴포넌트 생성
